Put all models test case in module
